Rename existing store
